Unconditionally register schema even for manual registration.
The general concept is that I want a centralized location where you
can find all of the registrations for a library.  I cannot do this
if I don't codegen all of the schemas in one spot--right now,
most schemas get generated, but not manually registered ones. Let us
assume that manual registration has to do with the actual
implementations; nothing strange is going on with the schema
definition itself.  Make it so.

Signed-off-by: Edward Z. Yang <ezyang@fb.com>

ghstack-source-id: 78f67cfcb8ae7243e16fb197350ea558dcdc3154
Pull Request resolved: https://github.com/pytorch/pytorch/pull/36250